+++ docs/conf.py
@@ -29,6 +29,9 @@ import ribs
 sys.path.insert(0, os.path.abspath(".."))  # Detect ribs.
 sys.path.append(os.path.abspath("./_ext"))  # Detect extensions.
 
+# Set canonical URL from the Read the Docs Domain
+html_baseurl = os.environ.get("READTHEDOCS_CANONICAL_URL", "")
+
 DEV_MODE = os.environ.get("DOCS_MODE", "regular") == "dev"
 READTHEDOCS_VERSION = os.environ.get("READTHEDOCS_VERSION", "stable")
 READTHEDOCS_LANGUAGE = os.environ.get("READTHEDOCS_LANGUAGE", "en")
@@ -135,6 +138,10 @@ html_logo = "_static/imgs/icon.svg"
 html_favicon = "_static/imgs/favicon.ico"
 html_title = (f"pyribs (stable - v{version})" if READTHEDOCS_VERSION == "stable"
               else f"pyribs ({READTHEDOCS_VERSION})")
+
+# Tell Jinja2 templates the build is running on Read the Docs
+if os.environ.get("READTHEDOCS", "") == "True":
+    html_context["READTHEDOCS"] = True
 
 # material theme options (see theme.conf for more information)
 html_theme_options = {